+++ src/debounce.js
@@ -3,14 +3,24 @@
 
 Bacon.EventStream.prototype.debounce = function(delay) {
   return withDesc(new Bacon.Desc(this, "debounce", [delay]), this.flatMapLatest(function(value) {
-    return Bacon.later(delay, value);
-  }));
-};
+    return Bacon.later(delay, value)
+  }))
+}
 
-Bacon.Property.prototype.debounce = function(delay) { return this.delayChanges(new Bacon.Desc(this, "debounce", [delay]), function(changes) { return changes.debounce(delay); }); };
+Bacon.Property.prototype.debounce = function(delay) { 
+  return this.delayChanges(new Bacon.Desc(this, "debounce", [delay]), function(changes) { 
+    return changes.debounce(delay)
+  })
+}
 
 Bacon.EventStream.prototype.debounceImmediate = function(delay) {
   return withDesc(new Bacon.Desc(this, "debounceImmediate", [delay]), this.flatMapFirst(function(value) {
     return Bacon.once(value).concat(Bacon.later(delay).filter(false));
   }));
 };
+
+Bacon.Property.prototype.debounceImmediate = function(delay) { 
+  return this.delayChanges(new Bacon.Desc(this, "debounceImmediate", [delay]), function(changes) { 
+    return changes.debounceImmediate(delay)
+  })
+}