+++ src/retry.js
@@ -6,7 +6,7 @@ Bacon.retry = function(options) {
   }
   var source = options.source;
   var retries = options.retries || 0;
-  var maxRetries = options.maxRetries || retries;
+  var retriesDone = 0
   var delay = options.delay || function() {
     return 0;
   };
@@ -21,7 +21,7 @@ Bacon.retry = function(options) {
       return source().endOnError().withHandler(function(event) {
         if (event.isError()) {
           error = event;
-          if (!(isRetryable(error.error) && retries > 0)) {
+          if (!(isRetryable(error.error) && (retries===0 || retriesDone < retries))) {
             finished = true;
             return this.push(event);
           }
@@ -40,10 +40,10 @@ Bacon.retry = function(options) {
     } else if (error) {
       var context = {
         error: error.error,
-        retriesDone: maxRetries - retries
+        retriesDone
       };
       var pause = Bacon.later(delay(context)).filter(false);
-      retries = retries - 1;
+      retriesDone++
       return pause.concat(Bacon.once().flatMap(valueStream));
     } else {
       return valueStream();